Fixed a problem with overwriting old/existing control schemes.

+++ PrimeHack Updater/Updater.cs
@@ -83,7 +83,8 @@ namespace PrimeHack_Updator
 
                 foreach (string file in files)
                 {
-                    File.Copy(file, docspath + file.Replace(".\\Profiles\\", ""), true);
+                    if (!File.Exists(file))
+                        File.Copy(file, docspath + file.Replace(".\\Profiles\\", ""), true);      
                 }
 
                 Directory.Delete(".\\Profiles\\", true);
